+++ main.py
@@ -1,6 +1,4 @@
 import os
-
-from pyrebase.pyrebase import Database
 os.environ['KIVY_IMAGE'] = 'sdl2'
 
 import kivy
@@ -12,37 +10,21 @@ Config.set('graphics', 'height', '812')
 
 from kivy.app import App
 from kivy.lang import Builder
-from kivy.uix.screenmanager import Screen
 
 from firebase import Authentication, DatabaseInteraction
 
 
-class WelcomeScreen(Screen):
-    pass
-
-class LoginScreen(Screen):
-    pass
-
-class HomeScreen(Screen):
-    pass
-
-class NewLogScreen(Screen):
-    pass
-
-
-GUI = Builder.load_file('main.kv')
-
 class MainApp(App):
-    
     def build(self):
         self.auth = Authentication()
         self.db = DatabaseInteraction()
 
-        return GUI
+        return Builder.load_file('main.kv')
 
     def change_screen(self, screen_name):
         screen_manager = self.root.ids['screen_manager']
         screen_manager.current = screen_name
+
 
 if __name__ == "__main__":
     MainApp().run()