+++ CHANGELOG.md
@@ -7,7 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Authentication system
 - Scrolling view on homescreen
 
 ## [0.1.0dev] (3/10/2020)
@@ -16,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Sign up and sign in screens.
 - Finished Home screen
+- Authentication system
 
 ## Changed